requirements.txt
redis
fastapi
google.generativeai
nltk
numpy
pydantic
redis
slowapi
uvicorn
setuptools
psycopg2-binary
psutil
torch
transformers
aiohttp
huggingface_hub